refactor console page into feature modules
